split up matcher.js
