+++ index.js
@@ -1,6 +1,6 @@
-import {defineRule, definePatternRule, canApplyRule, applyRule} from './lib/matcher'
+import {defineRule, definePatternRule, canApplyRule, applyRule} from './lib/rule'
 import {flattenOperands} from './lib/utils'
-import * as rules from './lib/rules.js'
+import * as rules from './lib/rule-list.js'
 
 export {
     applyRule,
