+++ lib/__test__/rule-list.test.js
@@ -1,8 +1,8 @@
 import assert from 'assert'
 import {parse, print, evaluate} from 'math-parser'
 
-import {applyRule, canApplyRule} from '../matcher.js'
-import * as rules from '../rules.js'
+import {applyRule, canApplyRule} from '../rule'
+import * as rules from '../rule-list'
 
 const applyRuleString = (rule, input) => print(applyRule(rule, parse(input)))
 const canApplyRuleString = (rule, input) => canApplyRule(rule, parse(input))
