+++ lib/matcher.js
@@ -3,23 +3,13 @@
  */
 import {build, query} from 'math-nodes'
 import {print} from 'math-parser'
-import evaluate from 'math-evaluator'
 import {replace, traverse} from 'math-traverse'
 
-import {
-    checkBounds,
-    clone,
-    fixMinuses,
-    getPlaceholders,
-    isVariableLengthPattern,
-    removeUnnecessaryParentheses,
-} from './utils'
+import {clone, getPlaceholders, isVariableLengthPattern} from './utils'
 
 const isArray = (val) => Array.isArray(val)
 const isObject = (val) => typeof val === 'object' && val !== null
 const isFunction = (val) => typeof val === 'function'
-
-const isPlaceholder = node => node && node.type === 'Placeholder'
 
 /**
  * Expand a variable length pattern.
@@ -41,7 +31,7 @@ const isPlaceholder = node => node && node.type === 'Placeholder'
  *
  *   #a_0 - #a_1 + #a_2
  */
-const _expandVariableLengthPatternNode = (pattern, length, negatives) => {
+export const _expandVariableLengthPatternNode = (pattern, length, negatives) => {
     const firstPattern = pattern.args[0]
     const firstPlaceholders = clone(getPlaceholders(firstPattern))
 
@@ -51,8 +41,8 @@ const _expandVariableLengthPatternNode = (pattern, length, negatives) => {
     const subscriptCounters = {}
 
     // Variable length patterns can be nested so we need to guard against
-    // expanding them more than once.  For an example, ee COMMON_DENOMINATOR
-    // in rules.js.
+    // expanding them more than once.  For an example, see COMMON_DENOMINATOR
+    // in rule-list.js.
     let skipExpansion = false
 
     for (let i = 0; i < length; i++) {
@@ -390,160 +380,4 @@ export const match = (matchFn, input, constraints = {}) => {
     })
 
     return result
-}
-
-/**
- * Replace any Placeholder nodes in an AST pattern with values from a
- * placeholder dictionary.
- *
- * This handles expansion of variable length patterns.
- */
-export const populatePattern = (pattern, placeholders) => {
-    const expandedPattern = replace(clone(pattern), {
-        enter(node) {
-            if (isVariableLengthPattern(node)) {
-                const firstPattern = node.args[0]
-                const firstPlaceholders = clone(getPlaceholders(firstPattern))
-
-                let variablePlaceholder = null
-                for (const [name, placeholder] of Object.entries(firstPlaceholders)) {
-                    if (placeholder.subscript) {
-                        variablePlaceholder = placeholder
-                        break
-                    }
-                }
-
-                const values = placeholders[variablePlaceholder.name]
-                const length = parseInt(Math.max(...Object.keys(values))) + 1
-
-                return _expandVariableLengthPatternNode(node, length, placeholders.negatives)
-            }
-        }
-    })
-
-    return replace(expandedPattern, {
-        leave(node) {
-            if (node.type === 'Apply' && isPlaceholder(node.op) && node.op.name === 'eval') {
-                return build.numberNode(evaluate(node.args[0]))
-            } else if (node.type === 'Placeholder' && node.name in placeholders) {
-                if (node.subscript) {
-                    const subscript = print(node.subscript)
-                    return clone(placeholders[node.name][subscript])
-                } else {
-                    return clone(placeholders[node.name])
-                }
-            }
-        }
-    })
-}
-
-/**
- * Convert pattern AST to matching function.
- *
- * The resulting function accepts an AST node.  Passing this to 'match' along
- * with the root node of an AST will will return the first node whose subtree
- * matches the structured described in the pattern.
- */
-export const patternToMatchFn = (pattern, constraints) =>
-    node => {
-        const result = matchNode(pattern, node, constraints)
-        if (result) {
-            return {
-                node,
-                ...result,
-            }
-        } else {
-            return null
-        }
-    }
-
-/**
- * Convert pattern AST to a rewrite function.
- *
- * The resulting function will accept an AST node and a placeholders object.
- * The node passed to the function is the matched node returned by 'match' in
- * applyRule.
- */
-export const patternToRewriteFn = pattern =>
-    (node, placeholders) => removeUnnecessaryParentheses(populatePattern(pattern, placeholders))
-
-
-// Public API
-
-/**
- * Define a rewrite rule.
- *
- * While this function provides more flexibility when defining rewrite rules,
- * it's harder to use.  Use definePatternRule whenever possible.
- *
- * @param {Function} matchFn a function taking an AST returning a match result
- * @param {Function} rewriteFn a function taking an AST and placeholders object
- * and returning an updated copy of the AST
- * @param {Object} [constraints] an optional dictionary with placholder names
- * for keys and functions for values.  The functions take an AST node and return
- * a boolean.
- */
-export const defineRule = (matchFn, rewriteFn, constraints = {}) => {
-    return {
-        matchFn,
-        rewriteFn,
-        constraints,
-    }
-}
-
-/**
- * Define a rewrite rule using patterns.
- *
- * @param {node} matchPattern pattern to match
- * @param {node} rewritePattern pattern to output
- * @param {Object} constraints object with placeholder names for keys and
- * functions taking nodes and return booleans as values
- */
-export const definePatternRule = (matchPattern, rewritePattern, constraints = {}) => {
-    // TODO: sanity checking for patterns being passed in
-    // - rewritePattern can't have any Pattern nodes with names not in matchPattern
-    const matchFn = patternToMatchFn(matchPattern, constraints)
-    const rewriteFn = patternToRewriteFn(rewritePattern)
-    return defineRule(matchFn, rewriteFn, constraints)
-}
-
-/**
- * Check if the rule can be applied before actually applying it.
- */
-export const canApplyRule = (rule, node) =>
-    !!match(rule.matchFn, node, rule.constraints)
-
-/**
- * Apply a rule to the given input pattern.
- *
- * Match a single node in input based on matchPattern.  If a match is found,
- * the node will be replaced with the rewritePattern with all of the pattern's
- * placeholder nodes replaced with the placeholder values determined when making
- * the initial match.
- */
-export const applyRule = (rule, input) => {
-    const {matchFn, rewriteFn, constraints} = rule
-    const {node, placeholders, indexes} = match(matchFn, input, constraints)
-    const matchedNode = node
-
-    if (matchedNode) {
-        const replacement = fixMinuses(rewriteFn(matchedNode, placeholders, indexes))
-        return removeUnnecessaryParentheses(replace(input, {
-            leave(node) {
-                if (node === matchedNode) {
-                    if (indexes && checkBounds(indexes, node.args)) {
-
-                        const {start, end} = indexes
-                        // TODO: make running that pass optional so that it
-                        // can be done separately if necessary
-                        node.args.splice(start, end - start, clone(replacement))
-                    } else {
-                        return clone(replacement)
-                    }
-                }
-            }
-        }))
-    } else {
-        return input
-    }
 }
