+++ lib/rule-list.js
@@ -4,7 +4,7 @@ import {gcd, lcm, nthRoot, primeFactorization, abs} from 'math-evaluator'
 import {build, query} from 'math-nodes'
 import {traverse} from 'math-traverse'
 
-import {defineRule, definePatternRule, applyRule, canApplyRule} from './matcher'
+import {defineRule, definePatternRule, applyRule, canApplyRule} from './rule'
 import {isPolynomialTerm, getCoefficient, getVariableFactors} from './rules/collect-like-terms.js'
 import {clone, getRanges} from './utils'
