+++ lib/rules/collect-like-terms.js
@@ -1,7 +1,8 @@
 import {parse, print} from 'math-parser'
 import {build, query} from 'math-nodes'
 
-import {canApplyRule, defineRule, populatePattern} from '../matcher'
+import {canApplyRule, defineRule} from '../rule'
+import {populatePattern} from '../pattern'
 import {flattenOperands} from '../utils'
 
 const clone = node => JSON.parse(JSON.stringify(node))
